requirements.txt
streamlit
streamlit-chat
openai
python-decouple